Added docs specification for structs and maps
The solidity style guide now includes instructions on how to document
structs and mappings. Othere Solidity projects seem to not do this, so
it was not possible to adhere to a community standard.
Instead, I compared how rust and golang recommend to do it and chose the
rust way of documenting structs and mappings.

In golang, a struct's field would be documented on the same line after
the name.

I used `/** */` instead of `///` as our style guide explicitly says to
avoid `///`.