Added the ability to report blocks to origin
See TODOs inline and please discuss.

The core contract is now actually tracking reported block headers at
heights. Different block headers can be reported at the same height, but
one header can only be reported once per height. A validator should
probably vote instead if the block has been reported before.

An event informs that a block was reported.

There is also a function for the validator to get all block hashes at a
given height. The validator can use that to check if the block has been
reported before reporting it. If multiple validators report the same
block at the same origin chain height, only one will be the reporter and
pay for the report. The others' transactions will be reverted.

I reduced the interface as some methods are not possible to be
implemented without committing blocks. They should be added when the
voting/committing is implemented.

I cleaned up the syntax and documentation.

I added a new check in the events decoder that will skip any event that
was not emitted by the contract that was passed to the decoder. This
change does not break any existing tests. The test that uses the event
decoder most likely is only interested by events that were emitted by
the provided contract.

Fixes #253