Reporting an auxiliary checkpoint to auxiliary
It is now possible to report an auxiliary checkpoint to auxiliary. The
report must be at a valid height in relation to the epoch length and a
checkpoint can only be reported once.

An event announces a successful report.

A mapping stores all reports.

The interface and config have been extended accordingly.

Fixes #265