Constructing the auxiliary stake contract
The contract is initialised on auxiliary with a starting set of
validators. The starting set consists of addresses and their deposit.
All starting validators are assumed to join the validator set at height
0 and to not have left the set of validators, yet. The stake must be
greater than zero, otherwise it is irrelevant.

The contract also tracks the total stakes at any given height so it is
known what the weight of a validator is at a given height in comparison
to the total stake.

Fixes #270