New validators are now added at new OSTblocks
The OSTblock gate contract is allowed to update the OSTblock on the
auxiliary stake contract. It increases the OSTblock height and adds the
new validators.

The address of the OSTblock gate is set on construction and cannot be
changed later.

It is currently not allowed to remove validators. That will be added in
a later step and requires stitching with forward and rear validator sets
like described in the Casper FFG paper.

Refactored adding of new validators into a private method as it is used
when updating an OSTblock as well as on construction.

Added JavaScript tests to assert the correct behaviour when initialising
an auxiliary stake contract.

Added Solidity tests to test the actual update method. In order to test
for reverts, I added a revert proxy as described in the truffle
documentation.

Fixes #282