Created interfaces for the core contracts
Changed some existing interfaces and added some new interfaces. Renamed
all interfaces and contracts to clarify their purpose.

The voter contract is now called "Polling Place". From Wikipedia:

> A polling place is where voters cast their ballots in elections.

There is a single polling place for both block stores. The address of
the block store is an argument to the vote function to clarify what
block store the vote is for.

A parent block store now exists in a rudimentary state. The auxiliary
block store and the origin block store are block stores. The block store
has its own BlockStoreInterface.

All missing methods that exist on the interfaces have been added to the
contracts that implement them with a `revert()` that the method is not
(yet) implemented.

I removed some TODOs and created GitHub issues instead to clean up the
code base.

The stake interface already includes all methods for validators to enter
and leave the validator set. We may want to remove the methods related
to ekiting the set of validators in order to squelch the warnings from
the compiler.

*Open Questions:*

1. Should we reduce the number of arguments to the `vote()` method on
the PollingPlace? An option is a `bytes _voteRlp` argument that is the
entire vote RLP encoded. In the method we would still need to get all
the variables for verification and storage.

2. The `proposeBlock()` method of the `OriginCoreInterface` now has an
argument `_dynasty`. Instead, I think it would make sense to add the
(auxiliary) dynasty to the OSTblock header.

3. Two questions on the vote verification method of the
`OriginCoreInterface`:

3a. How do we link votes (auxiliary FFG) to OSTblocks? Based on the
state root? Then we would need to add the state root to our vote object.
I think a good idea would be to add the block hash of the finalised
auxiliary checkpoint to the OSTblock header.

3b. How are the votes proven? Where is the trie of votes kept? Only in
the rum nodes' memory?

Fixes #295