Stake tracks and returns weights
The stake contract now returns the weight for a specific validator at a
spcific height as well as the total weight at any given height.

In order to easier track the weight, I added an array to track all
deposited addresses and use that to iterate over the mapping of
validators. By mapping them from their address it is easier to access
them in general for any methad.

I added a private method to get the validator weight that implements the
external function. Since interfaces will only allow external methods but
I wanted to re-use the functionality in another function, I assumed this
is the way to go now.