Implementation of the PollingPlace
All methods of the PollingPlace have been changed and the polling place
interface as well as the block store interface have been changed.

I created a mock of the block store in `contracts/test/core` to mock the
block store dependency of the polling place.

I added a new dependency to `package.json`. `ethereumjs-abi` is required
to do packed abi encoded signing (like solidity does) from within the
javascript unit tests.

There is a new requirement that the target height of a vote must be
higher than the "core height". This is in line with casper where the
target must be within the current dynasty. However, this could
potentially open up a new attack vector. It is now possible to vote on
targets that are currently considered within the current meta-block, but
will move into the next meta-block if the current one is closed before
the target blokc.

Fixes #322