Ported repository to Truffle v5
Most important changes:
- web3 is now version 1.0. Some method calls had to be updated.
- BigNumber has been replaced with bn.js.

The changes lead to a number of changes in the test files.

One thing is not clear to me. It could be a bug in the test or the
contract:
In `OpenSTUtility->ProcessRedeeming->STPrime->successfully processs` I
had to change one of the tests. It is **now** checked that the balance
of STPrime **after** the processing has **increased**. Before it was
tested that it decreased. I checked the code and `value()` is sent to
STPrime. So either the old test (`minus` instead of `add`) was faulty or
there is a bug in the contract code. **Please Check!!**

See also the truffle release notes: https://github.com/trufflesuite/truffle/releases/tag/v5.0.0-beta.0