Added web3 requirement to every test file
We wanted to add web3 from package so that we can update the web3
version indepentdently from the version shipped with truffle. To have a
single place where to change and setup web 3, a new lib file was added
that must from now on be required at the top of every test file.