Implementation of the BlockStore
The block store stores headers of blocks. Not all headers have to be
reported. It is only required to report all headers that should become
justified checkpoints.
The block store tracks all justifications and finalisations through
calls to the `justify()` method. Only the polling place can call that
method.

Tests are written in one file per method that is tested.