Implementation of the AuxiliaryBlockStore
The auxiliary block store is a specialised version of a block store. In
addition to tracking reported blocks, it only accepts blocks whose
parents have been reported before. It also tracks the accumulated
transaction root and the accumulated gas.

This commit also fixes a bug in the BlockStore. It is from now on not
possible anymore to have valid votes/justifications with a target
checkpoint that was previously justified with a different source.

All methods are tested and all cases shuld be covered. I created a
`helpers` directory in the test directory for the most common tasks and
data. A possible improvement could be differing state roots and
transaction roots for the blocks in the shared data.

Fixes #356