Stake now has an initial set of validators
The initial set of validators cannot be added upon construction. It is
not possible to transfer token to the contcract during construction, as
`address(this)` is not available. Giving its own address as a
constructor parameter is possible, but error prone.

Instead, a second `initialize()` function was added. It can only be
called once (successfully) and it initializes the initial validator set
at height `0` with the given values.

Furthermore, there exists a new modifier that checks on crucial methods
that the total weight of the validators is above a minimum weight that
is defined during construction.

According to the new framework with a directory per contract and a file
per tested method.

Moved shared functions into a utils class.

Replaced bignumber with bn.