Auxiliary block store now uses correct transition object
The auxiliary transition object was not used anywhere in the code. It is
now used wherever we need the auxiliary transition object. Mostly the
auxiliary block store.

In order to be able to calculate the transition object and hash at an
auxiliary checkpoint, we need to know the origin dynasty and latest
finalized block hash. Those two fields are now tracked on the auxiliary
block store per reponted block by calling on the origin block store. The
address of the origin block store was added to the constructor.

To decide whether a vote is valid, the auxiliary block store now
verifies the given transition hash against the hash that it calculates
itself based on the data that is available  at the source checkpoint. To
do that, `isValidTransitionHash` has been made internal on the
`BlockStore` and is overridden by the `AuxiliaryBlockStore`.

Added `getHead()` and `getCurrentDynasty()` to the `BlockStoreInterface`
and `BlockStore` and `BlockStoreMock`. The auxiliary block store uses
these functions to get the latest observations of the origin block store
when it stores a block report.

I updated the auxiliary block store tests to check for the corret
transition hash. Furthermore I added the accumulated values to the
shared data object.

I added solidity tests for the `MetaBlock` library.

All core identifier types were changed to `bytes20` which is the current
implementation. Must be changed to `bytes32` across the whole project
coordinated.

I tried caching only the transition hash per reported block in the
auxiliary block store instead of all the individual data points like
accumulated gas and accumulated transaction root. It turned out
impossible to do, as the dynasty of a block can change when an earlier
block gets finalized. Thus, the transition hash cannot be cached at the
time of a block report.

Fixes #336