Added changelog for the changes so far in v0.10
I added things that I am aware of or found in the closed pull-requests.
However, I am pretty sure there is more stuff that happened on the
gateway side of things that I am not necessarily aware of. So please
check and suggest any and all additions that you think should be part of
the change log!