Message and Intents are now hashed as per EIP712
The Message, StakeIntent, and RedeemIntent now have a new structure and
different fields that are part of the hash. Stake and redeem intent now
consist of amount, beneficiary, and (originating) gateway address. The
Message has as first parameter a hash which is not specific to staking
or redeeming.
All three pre-images now include a type hash as per EIP712.

This means that I was able to remove a lot of stake/redeem specific code
for message hashing. A message's hash is now generated regardless of the
kind of intent it holds.

I added a method `storeMessage` which creates a message hash and stores
the message at the hash in the mapping. The reason was to circumvent a
"stack too deep" error which prevented me from creating the hash
directly in confirm stake/redeem intent.

I restructured the tests in an effort to make them more easily
understandable. I spent quite some time wrapping my head around the
construct.

I did not add the domain separator yet, as the hashes are never signed
or the signature proven/verified.

:warning: Please check that the new logic of how and what is hashed
still guarantees correctness and prevents fraud when relaying the
messages :warning:

Fixes #415