Tests assert that the gateway nonce increases by 1
I extended the `stake.js` and `redeem.js` tests to include an additional
testcase. It asserts that the nonce increases by 1 when stake or redeem
is called on the (co)gateway.

The alternative to put the test cases into `get_nonce.js` wasn't
feasible. Extracting the stake/redeem functionality from `stake.js` and
`redeem.js` would be very tedious and time consuming. Instead, I added a
comment to the `get_nonce.js` file that the tests about increasing the
nonce are part of the stake and redeem tests.

I updated the affected files to pass linting.

Fixes #625