Unified JS code style
I added a prettier to format the code.
I formatted the code with the prettier.

We have more than 700 linting issues left, which I cannot fix all in
this PR. Most of them are max-line and prefer-destructuring.
I suggest we start fixing the files one at a time when we work in them.

Fixes #404